Adds chunks support
This is some preliminary work to add chunk loading, ie. "random access" on chain.

There is still a lot to refactor, and I just discovered that when loading high blocks there is some strange behavior.

This change make the assumption that the chain always starts at block 0, and then moves to the desired place. This makes things easier to compute coordinates.